Addessables and Bootstrap Scene flow
Latest Addressables package )(v 1.21.1)
Non-addressable Bootstrap scene as only scene in BuildSettings
Bootstrap loads PersistentManagers, which is an addressable scene.
PersistentManagers has an AppController object that picks which game scene to load
For now Game scene is either AccountDemo or a dummy TitleScreen and the switch is a preprocessor define (ACCOUNT_DEMO).