Unity map update to display the new State object
`PluginController.cs` has been pretty much gutted as most of its code was handling GraphQL and the private key / session key siginin stuff. Now all that is handled with `DawnSeekerBridge` or the shell, all it needs to do now is deserialise the new state objects, send events and handle action dispatch requests.

Action dispatching is yet to be implemented