qol: allow make dev with stale map build
unity is fussy.

it will complain if you try to do a cli build while a unity editor is open.

in the situation where you have built the map at least once, and don't
care that the map build is going to fail right now, make dev will now
prompt and ask if you want to continue with your old stale map anyway to
avoid having to go close unity just to get the frontend up