remove REGISTER_BUILDING_PLUGIN
all client plugins now registered via the same REGISTER_CLIENT_PLUGIN

ClientPlugin plugin nodes indicate their support for other nodes via the
"Supports" relationship.

This allows for one plugin to support many things.

Right now there are artifical restrictions that:

* the registrant owns the target node they are Support-ing
* the target node is of kind BuildingKind or Null

this restriction is only in place as it is not clear how or if we want
to allow multiple plugins for the same thing