Map state restructure
- Regenerated C# types however they are  a mess at the moment

TileHelper can return the position of a tile again

And this was horrible. Locations on Seekers haven't had their properties codegenned so I had to fish them out as generic Newtonsoft JObjects. I'm going to have to fix this properly at the TypeScript end but not sure how to yet so will crack on with this hack

Selection working again

- Intent.NONE set to be `null` instead of empty string
- Manually edited the generated Schema to allow for buildings to be null

Player seeker displaying in the right place

- I'm rendering the seeker using the seeker found in the player object instead of checking if the seeker on the tile belongs to the player or not. This at least smooths over the problem of the map state being one behind where the seeker currently is but doesn't fix why it's like that.