C# state schema regen and session check
Session intent now checks the newly exposed sessions array on a tile to check if there is an active session or not. The tile won't light up as valid if there is currently a combat session in play.
The model annotation has now accessed in C# due to the schema regen