BUG: Actions occurring in the same block were being overwritten
Because the events use annotations which are indexed by label all events that happen in the same block have the same label which meant that only the last event survived. An example of this would be if a unit moved from one side of combat to the other. The leave and join events happen in the same call and only join was indexed which meant that we saw the same unit on both sides of the battle and it wouldn't finalise because the hashes wouldn't match