Incrementing a counter after each session update
As GraphQL doesn't give me the list of events in the order they happen I rely on the block number to order them. If more than one event happens within the same block there isn't any guarantee we are ordering the list correctly which means it would sometimes work and sometimes not.