Remove of SeekerController from EnemyPrefab
The enemy prefab had the `SeekerController` component added to it which would throw an error on update as the enemy prefab isn't constructed in the same way. Replaced with the `MapElementController` same as buildings and now it doesn't fail during render. This has the side effect of making player highlights work again