Block time updates happen every second and combat actions count for the current block
By increasing the time we check for block time update it solves the problem where we had a block time of 0 for the first 10 seconds of play
Combat logic now counts actions that occur on the current block we are on. The state calculation was exiting early if the action was occurring on `>= currentBlock`. This meant for the first tick of combat there was no state