hide BuildingKinds without "building" annotation
hide any BuildingKinds from the construction list that have a "model"
annotation that is not "building".

this allows paul to annotate any buildings he doesn't want visible in
the list as "story-building" or something like that.

this is a temporary patch for Blueprint or until we have a proper
solution for the constructable list