Bags show up on defeated tiles if the player has a reward
We check if the tile has a session and if the player has any rewards associated with that session.
Known issue is that we don't know without calculating the full combat state which side won so for now I'm going to assume that if there are rewards that it would have been the defenders that had lost. Bags belong to the session not the tile so in reality the bag could be shown on either tile