tilefab: a tool for generating tile manifests
we wanted a way to quickly draw out some map shapes and place some
starter buildings for playtests. it is a bit of a faff writing manifests
for the locations of things by hand.

this renders a big hexagon of a defined diameter, and lets you "paint"
discovered tiles on it to create nice shapes.

you can also load BuidlingKind manifests and place Buildings at
locations on the map.

a "preview" renders the tiles with the unity map to get a better feel
for the result.

export the manifest yaml and place it in the fixtures dir to have it
deployed with the game.