Only the owner of the extractor is allowed to extract goo from it
Only the player that built the extractor can extract goo from it. This is both enforced at a frontend level where a message is displayed if the a player that doesn't own the building tries to use it.
This only applies to the extractors that were built by a player, extractors that are spawned via the map.yaml are public.

On a technical level it's actually the owner of the output bag that determines who can use the building or not. Done this way so that we can still have public extractor if we choose to also after an extractor has been built, it would be possible for a player to set the bag to public if they so wished

I'm also spawning mobile units with 100 of each resource so they can build an extractor (as per sprint goal)