deployBuilding Quest can specify specific recipe  items
deployBuilding Quest task can now define a required input and/or output item
craftInput and craftOutput are optional task fields in the quest manifest if kind is deployBuilding.
They are strings that need to match the name of an existing item.
In contract state, these are set as balance-of-item relationships on the task node, with inputs having rel key 0 and outputs 1
In Quest graphql data these are retrieved as an array of key, item objects, which are used to check the agains the currently deployed player buildings to see if the input/output items specified exist in the buildingkind receipe slots.