Comparing weights instead of rates to determine if a tile has some or lots of goo
# What

Comparing weights instead of rates to determine if a tile has some or lots of goo. By comparing weights we are applying the same logic that we use to display either a small or big goo texture.

# Why

This solves the problem where a tile without a goo texture could say it had some goo on it. Even though this was technically true it didn't match the visuals