Teams now only exist in building storage
Lots of further modifications were done to optimise and refine the other functions, otherwise I was getting this error:
"Warning: Contract code size is 25281 bytes and exceeds 24576 bytes (a limit introduced in Spurious Dragon). This contract may not be deployable on Mainnet. Consider enabling the optimizer (with a low "runs" value!), turning off revert strings, or using libraries."

which was stopping me from deploying the building.