Added env vars to Make dev command
For example: `make dev MAP=tiny`
or `make dev MAP=performance-test ARENAS=3` to create 3 duck burger arenas with unique plugins for each.
`make dev` continues to behave like usual.

Instead of a new make command, when defining `MAP=performance-test`, if checks for that in .devstartup.js and executes a script to build the performance test from the scripts folder.