Removes contracts cache in `make dev`
This is because if a path that no longer exists is referenced in the solidity-files-cache.json file it would return an error.

This fix works for now but adds ~30 seconds to build time. A possible better fix I will experiment with next is to delete the problematic path from the json cache when the test-plugins folder is deleted.