ensure plugins are never executed in parallel
plugins were executing concurrently during their async operations, this
led to some strange behaviour and was unexpected.

this ensures that plugins always execute serially, and in the exact
expected order.

to reduce the chance of them stacking up in a queue, a timeout
is added for max plugin execution time