Fixed most active client having to be open to finalise
Now it seems to try to finalise combat on the 1st and 2nd most active clients. However, now with both 1st and 2nd most active clients closed, the 3rd client DOES try to finalise combat - which was broken in the previous commit.

Probably still needs thorough testing before it goes in main.