redstone option to deploy
add new option to the deployment Github Action to pick between "anvil"
and "redstone-holesky" chains to deploy to.

when redstone-holesky is selected, it will deploy the contracts using
the REDSTONE_DEPLOYER_PRIVATE_KEY stored in github secrets, and setup the
indexer/sequencer to use the REDSTONE_SEQUENCER_PRIVATE_KEY.

ds apply of the selected map assets is done by globalally installing the
latest cli in the github action. so it is possible this could be out of
sync with the branch you are deploying.... for this reason, we should
only really use this option with `main` branch