single deployment, multi zone
allow players claim a "zone" (collection of tiles, buildings, bags,
units) and have full dev powers for executing DEV_* actions against
entities within that zone.

zone ownership is managed by an ERC721 NFT contract. ownership of the
NFT gives ownership of the zone. Ownership can be transfered by
transferring the NFT.

There are 32k available zones (a limitation that is really a side effect
of the coord system which is an int16).

Since zones are a limited resource, there is a mint price for claiming
a zone. this is in place as an anti abuse mechanism, since someone
claimning them all would brick the deployment.

The "front page" of the site is now a list of zones and way to claim
your zone.

As there are many zone and a single deployment, the ds-cli now requires
an additional -z flag to pass your zone id.

Manifests are NOT zone-aware, during apply any zone args are filled in
with the -z value.

GraphQL queries have been updated to split things into "global" and
"zone" sets of things.

Global thing are shared between all zones. Things like BuildingKinds,
ItemKinds, Plugins, Players

Zone things are found by looking at the "Parent" edge and are local to
the zone. Things like Tiles, CombatSessions, Buildings.

Unit location detemins which zone the unit is in. So the MOVE
actions now handle setting the zone parent.

To make local development close to the current setup zone-1 is always
claimed by the deployer key, and there is now a new "local-dev-key"
signin option to make it easy to signing with an account that has enough
ETH to mint your own zone. The main difference should be the additional
need to pass -z during ds-cli commands