Fixed DvB by removing option to choose team buiildings
The world.buildingkinds object has gone and DvB was using this to give players a choice of of the building that would be counted for each team.
These are now hardcoded instead.

To make the game flexible again I think the plugins would need access to global state - not sure how hard that is to provide or maybe they already do have access.