Support applying named foundry compiled artificats directly
This has the benefit of exporting the complexities of compiling solidity
entirely to the foundry tool chain. And it offers clean integraiton with
external projects which are already setup to use it.

Does nothing unless the manifest directory contains a foundry.toml

If foundry.toml is present, source.file is interpreted as a contract name.
FOUNDRY_OUT is read from env and defaults to 'out'. An attempt is then made to
locate the compiled artifact under the resulting directory according to the
output file schema for foundry.

This could be improved by actually reading foundry.toml to get the out dir

If the derived file doesn't exist, fall back to the previous solc

Example remappings if your manifest dir is a sibling of a ds clone

```
@ds/=../ds/contracts/src/
cog/=../ds/contracts/lib/cog/contracts/src/
```

allow_paths will be needed in foundry.toml. A simple example:

```
[profile.default]
solc_version = "0.8.13"
verbosity = 4
allow_paths = ["../ds/contracts"]
```