Allow Buildings to destroy themselves
# What

Currently only the owner of the zone can destroy buildings. This PR allows the building kind implementation contract to destroy any instances of its building kind.

# Why

There are games that need to be able to destroy things either as part of the gameplay or as part of a game reset process.