Refactor combat system to only have one session per tile pair
The combat session only exists until it is finalised. Once finalised it is destroyed
As we are are no longer keeping historical sessions, rewards are transferred directly to the player's wallet. This also means that there are no 'reward bags' on the map.

The reason for this is that we were keeping an ever growing amount of combat sessions which caused issues on the frontend where we were fetching all of them