expose item data to client
ItemKind contracts can SET_DATA_ITEM just like BuildingKind contracts
can SET_DATA_BUILDING to let them store data on-chain and read back in
the client.... however we were not previously exposing the item data to
the plugins so it was useless.

this exposes the item data so that plugins can do interesting stuff with
it.

it updates the cocktail hut so that so "sips" are no longer just a
client-local thing and are now a globally shared on-chain counter of all
sips ever by everyone.

somewhat hilariously though, there's a bug in the implementation of
ITEM_USE that means you can only "use" items that are in either
bag-0-slot-0 or bag-1-slot-1 :joy: ... so that's a bit annoying