disable construct intent for buggy zones
there's a contract bug preventing the CONSTRUCT_BUILDING_MOBILE_UNIT
action for working as intended when the zone id is 40 or more.

as a temporary measure we are disabling the build intent for the
affected zones so that it reduces the potential confusion of either the
resulting "out of bounds" error or the silent failing due to these
errors only being debug visisble.